pyproject.toml
[tool.poetry]
name = "snex"
version = "2021.10.26"
description = "snex - snippet extractor"
authors = ["Joachim Bargsten <jw@bargsten.org>"]
homepage = "https://github.com/jwbargsten/snex"
readme = "README.md"
repository = "https://github.com/jwbargsten/snex"
documentation = "https://github.com/jwbargsten/snex"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
click = "^7.1.2"
pyhocon = "^0.3.57"
pyyaml = "^5.4.1"
requests = "^2.26.0"
chevron = "^0.14.0"
Jinja2 = "^3.0.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
coverage = {extras = ["toml"], version = "^5.4"}
pytest-cov = "^2.11.1"

[tool.coverage.report]
show_missing = true

[tool.coverage.paths]
source = ["src"]

[tool.coverage.run]
branch = true
source = ["snex"]

[tool.poetry.scripts]
snex = "snex.cli:main"
[tool.black]
line-length = 110
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
